backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
Pillow==10.4.0
requests==2.32.3
numpy==2.1.2
pyserial==3.5

